Prayer times are now calculated and shown to the user
The User location (latitude & longtitude) is now being retrieved using the geolocator package.

The User is now asked, when opening the prayer time screen, to give permission for the location to be used

Todays prayer time is being determined using the adhan package.

The Times are then shown in a table

close #1